Rename columns for course completion events, alter columns in csv export
