+++ src/main/java/uk/gov/cshr/report/domain/report/CourseCompletionReportRequest.java
@@ -1,7 +1,10 @@
-package uk.gov.cshr.report.domain;
+package uk.gov.cshr.report.domain.report;
+
 import io.hypersistence.utils.hibernate.type.array.ListArrayType;
 import jakarta.persistence.*;
-import lombok.*;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
 import org.hibernate.annotations.Type;
 
 import java.time.ZonedDateTime;
