+++ src/main/java/uk/gov/cshr/report/domain/report/CourseCompletionReportRequestStatus.java
@@ -1,4 +1,4 @@
-package uk.gov.cshr.report.domain;
+package uk.gov.cshr.report.domain.report;
 
 public enum CourseCompletionReportRequestStatus {
     REQUESTED,
