+++ src/main/java/uk/gov/cshr/report/service/messaging/coursecompletions/CourseCompletionMessage.java
@@ -19,11 +19,11 @@ public final class CourseCompletionMessage implements IMessageMetadata, Serializ
     private final String courseId;
     private final String courseTitle;
     private final Integer organisationId;
-    private final String organisationAbbreviation;
+    private final String organisationName;
     private final Integer professionId;
     private final String professionName;
     private final Integer gradeId;
-    private final String gradeCode;
+    private final String gradeName;
 
     @Override
     public String getQueue() {
