+++ src/main/resources/db/migration/postgresql/V2.4__amend-course-completion-event-columns.sql
@@ -1,0 +1,5 @@
+ALTER TABLE course_completion_events
+    RENAME COLUMN grade_code TO grade_name;
+
+ALTER TABLE course_completion_events
+    RENAME COLUMN organisation_abbreviation TO organisation_name;
